+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This release contains new migrations and requires to run the `migrate` command a
 
 ## Fixed
 
++ [#258](https://github.com/luyadev/luya-module-admin/issues/258) NgRest Crud search with pagination problem fixed. 
++ [#226](https://github.com/luyadev/luya-module-admin/issues/226) Fixed search indicator
 + [#267](https://github.com/luyadev/luya-module-admin/pull/267) I18n::decodeFindActive returned empty value for explicitly selected lang
 
 ## Changed